+++ modules/farm_rothamsted_researcher/farm_rothamsted_researcher.post_update.php
@@ -5,6 +5,7 @@
  * Update hooks for farm_rothamsted.module.
  */
 
+use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\views\Entity\View;
 use Symfony\Component\Yaml\Yaml;
 
@@ -31,4 +32,32 @@ function farm_rothamsted_researcher_post_update_2_14_remove_orcid_prefix(&$sandb
       ->condition('orcid', 'https://orcid.org/%', 'LIKE')
       ->execute();
   }
+}
+
+/**
+ * Add fields to proposal entity.
+ */
+function farm_rothamsted_researcher_post_update_2_16_notification_field(&$sandbox = NULL) {
+
+  // Create notification_enabled field.
+  $field_definition = BaseFieldDefinition::create('boolean')
+    ->setLabel('Notifications')
+    ->setDefaultValue(TRUE)
+    ->setRevisionable(TRUE);
+  \Drupal::entityDefinitionUpdateManager()->installFieldStorageDefinition(
+    'notification_enabled',
+    'rothamsted_researcher',
+    'farm_rothamsted_researcher',
+    $field_definition,
+  );
+
+  // Set default value for existing researchers.
+  \Drupal::database()->update('rothamsted_researcher_data')
+    ->isNull('notification_enabled')
+    ->fields(['notification_enabled' => TRUE])
+    ->execute();
+  \Drupal::database()->update('rothamsted_researcher_field_revision')
+    ->isNull('notification_enabled')
+    ->fields(['notification_enabled' => TRUE])
+    ->execute();
 }